Edit the HTML file with CSS
Created an external style sheet for one of the elements of the page, linked it to index.html and made the required design changes indicated in the README-file.
Internally styled the second element using CSS within HTML with the indicated colours and sizes.
Styled the button with background colour and font-size in-line.